refactor: Replace avatars with a common custom avatar component
