+++ src/components/user-avatar/UserAvatar.tsx
@@ -7,17 +7,9 @@ import {
   DropdownMenuSeparator,
 } from "@/components/ui/dropdown-menu";
 import { serverSideFetcher } from "@/lib/utils";
-import { Avatar, AvatarImage, AvatarFallback } from "@/components/ui/avatar";
 import { cn } from "@/lib/utils";
-
-type User = {
-  id: number;
-  email: string;
-  nickname: string;
-  profileImageUrl: string;
-  createdAt: Date;
-  updatedAt: Date;
-};
+import CustomAvatar from "../custom-avatar/CustomAvatar";
+import { Avatar } from "@/components/ui/avatar";
 
 async function getUserInfo() {
   const response = await serverSideFetcher(
@@ -26,47 +18,21 @@ async function getUserInfo() {
   const data = await response?.json();
   return data;
 }
-function getFirstCharacter(str: string) {
-  if (str.length > 0) {
-    return str.charAt(0);
-  } else {
-    return "";
-  }
-}
-
-const getBackgroundClass = (userId: number) => {
-  const classes = [
-    "bg-[#ffc85a]",
-    "bg-[#fdd446]",
-    "bg-[#9dd7ed]",
-    "bg-[#c4b1a2]",
-    "bg-[#a3c4a2]",
-    "bg-[#034694]",
-    "bg-[#e876ea]",
-    "bg-[#000]",
-    "bg-[#c3102b]",
-    "bg-[#7AC555]",
-  ];
-  return classes[userId % 10];
-};
 
 export default async function UserAvatar({}) {
   const userData = await getUserInfo();
-  const fallback = getFirstCharacter(userData.nickname);
 
   return (
     <DropdownMenu>
       <DropdownMenuTrigger asChild className="outline-none">
         <button className="flex gap-3 pl-6 items-center border-l border-[#d9d9d9]">
-          <Avatar className="rounded">
-            <AvatarImage src={userData.profileImgUrl} width={38} height={38} />
-            <AvatarFallback
-              className={`text-white font-semibold ${getBackgroundClass(
-                userData.id,
-              )}`}
-            >
-              {fallback}
-            </AvatarFallback>
+          <Avatar>
+            <CustomAvatar
+              imgUrl={userData.profileImageUrl}
+              nickname={userData.nickname}
+              userId={userData.id}
+              size={38}
+            />
           </Avatar>
           <span className="font-medium">{userData.nickname}</span>
         </button>
